requirements.txt
asgiref==3.11.1
certifi==2026.2.25
distlib==0.4.0
Django==6.0.4
django-cors-headers==4.9.0
django-jazzmin==3.0.4
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.15
filelock==3.25.2
gunicorn==25.3.0
inflection==0.5.1
packaging==26.0
pipenv==2026.5.2
platformdirs==4.9.6
Pillow
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-decouple==3.8
python-discovery==1.2.2
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
setuptools==82.0.1
sqlparse==0.5.5
tzdata==2026.1
uritemplate==4.2.0
virtualenv==21.2.1
whitenoise==6.12.0
